+++ src/utils.js
@@ -1,8 +1,3 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 const MonthFormat = {
   0: `January`,
   1: `February`,
@@ -42,24 +37,6 @@ const formatDateWithSlash = (date) => {
   return `${year}/${month}/${day}`;
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
 const getSortingFilms = (films, selectionParameter) => {
   return films.slice()
   .sort((left, right) => (right[selectionParameter] - left[selectionParameter]));
@@ -75,9 +52,6 @@ const onEscPress = (evt, action) => {
 
 export {
   onEscPress,
-  RenderPosition,
-  render,
-  createElement,
   formatDateToString,
   formatDateWithSlash,
   formatTime,
